feat(recipes): add recipes feature

+++ server/src/main/kotlin/org/nikol/roasti/features/likes/LikeRepository.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.jetbrains.exposed.v1.core.and
 import org.jetbrains.exposed.v1.core.eq
+import org.jetbrains.exposed.v1.core.inList
 import org.jetbrains.exposed.v1.jdbc.deleteWhere
 import org.jetbrains.exposed.v1.jdbc.insert
 import org.jetbrains.exposed.v1.jdbc.selectAll
@@ -68,7 +69,7 @@ class LikeRepositoryImpl : LikeRepository {
             transaction {
                 val rows = LikeTable.selectAll()
                     .where { (LikeTable.targetId eq targetId) and (LikeTable.targetType eq targetType) }
-                val count = rows.count()
+                val count = rows.count().toInt()
                 val isLiked = userId != null && rows.any { it[LikeTable.userId] == userId.value }
                 LikeInfo(isLiked, count)
             }
